% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmium_installed.R
\name{assert_tool_is_installed}
\alias{assert_tool_is_installed}
\title{Assert command line tool is installed}
\usage{
assert_tool_is_installed(tool)
}
\value{
Throws an error if the tool is not installed, invisibly returns \code{TRUE}
otherwise.
}
\description{
Asserts that a tool is installed, throwing an error otherwise.
Shamelessly stolen from rosmium
}
\keyword{internal}
