% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_geom.R
\name{construct_geom}
\alias{construct_geom}
\title{construct_geom}
\usage{
construct_geom(cpprnet_dt, complete = FALSE, cores = 1)
}
\arguments{
\item{cpprnet_dt}{a data table obtained with cppRosm extract_data()}

\item{complete}{bool. TRUE to reconstruct Polygons, else use just points.}

\item{cores}{integer. Run the computation on multiple cores using \code{parallel}.}
}
\value{
Modifies the provided cpprnet_dt object and adds an \code{sf} geometry column to it, with either full geometries or just the centroids.
Full geometries typically contain polygons and points.
}
\description{
From a cppRosm data table, make the geometries.
}
\examples{
cpprnet_filepath <- system.file(package = 'cppRosm','extdata','monaco-latest.osm.pbf')
cpprnet_dt <- cppRosm::extract_data(filename = cpprnet_filepath)

cppRosm::construct_geom(cpprnet_dt,complete=TRUE)

head(cpprnet_dt$geometry)

}
