% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppRnet-package.R
\docType{package}
\name{cppRnet-package}
\alias{cppRnet}
\alias{cppRnet-package}
\title{cppRnet: Fast en efficient OSM data in R}
\description{
This package provides a set of R function for working with OpenStreetMap extract data through the libosmium library in C++.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ischlo/cppRnet}
  \item Report bugs at \url{https://github.com/ischlo/cppRnet/issues}
}

}
\author{
\strong{Maintainer}: Ivann Schlosser \email{schlosserivann@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
