% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_graph.R
\name{extract_data}
\alias{extract_data}
\title{Extracting data from raw osm.}
\usage{
extract_data(filename)
}
\arguments{
\item{filename}{a local osm extract file with .osm or .osm.pbf extension.

local_osm_filename <- system.file("extdata",'map.osm', package = "cppRnet")

cppRnet::extract_data(local_osm_filename)}
}
\description{
Getting data from raw osm files.
}
