% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_graph.R
\name{extract_graph}
\alias{extract_graph}
\title{Extracting road graphs from OSM}
\usage{
extract_graph(filename, out = NULL)
}
\arguments{
\item{filename}{a local osm extract file with .osm or .osm.pbf extension.}

\item{out}{a local osm extract file with .osm or .osm.pbf extension.}
}
\value{
this function returns the value 1 to the R environment when it's completed.
It writes out 2 files containing nodes and segments. The files are written in the current directory and are called:
\code{nodes.csv},\code{road_segments.csv}.
}
\description{
This function helps extract a road network graph in a simple and light weight format.
In the background, it uses the power of the osmium library in C++, and copies the data in graph format locally as 2 .csv
files, one with nodes and their coordinates, with with segments and optional values.
}
