% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sec.R
\name{filter_sec}
\alias{filter_sec}
\title{Filtering secondary tags}
\usage{
filter_sec(data, keys, cores = 1)
}
\arguments{
\item{data}{a cppRnet data set, with an \code{attrs} column.}

\item{keys}{either a vector of valid key parameters, or a named list where names correspond to keys and
attributes are vectors of matching values.}

\item{cores}{number of cores to parallelise the computation}
}
\description{
This function provides a handy way of filtering a cppRnet data set for specific secondary tags.
It will return a data set to which will be added new columns based on the specific secondary keys and values that
are looked for.
}
\examples{

test_file <- system.file(package = 'cppRnet','extdata','map.osm')
data <- cppRnet::extract_data(test_file)


keys <- list("cuisine"=c("pizza","japanese")
            ,"takeaway"=c("yes"))

data_filtered <- cppRnet::filter_sec(data,keys)

}
