% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppRnet-package.R
\docType{package}
\name{cppRnet-package}
\alias{cppRnet}
\alias{cppRnet-package}
\title{cppRnet: Fast en Efficient OSM Data in R}
\description{
A set of fast and user friendly functions for working with large OpenStreetMap extract data. Developed on top of the C++ libosmium library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ischlo/cppRnet}
  \item Report bugs at \url{https://github.com/ischlo/cppRnet/issues}
}

}
\author{
\strong{Maintainer}: Ivann Schlosser \email{schlosserivann@gmail.com} (\href{https://orcid.org/0009-0004-4099-3198}{ORCID})

}
\keyword{internal}
