% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_graph.R
\name{extract_graph}
\alias{extract_graph}
\title{Extracting road graphs from OSM}
\usage{
extract_graph(filename)
}
\arguments{
\item{filename}{a local osm extract file with .osm or .osm.pbf extension.

local_osm_filename <- system.file("extdata",'map.osm', package = "cppRnet")

cppRnet::extract_graph(local_osm_filename)}
}
\description{
This function helps extract a road network graph in a simple and light weight format.
In the background, it uses the power of the osmium library in C++, and copies the data iin graph format locally as 2 .csv
files, one with nodes and their coordinates, with with segments and optional values.
}
